Initial check in of tboot recipe.
Currently builds just the tboot binary.  OE and tboot do some
serious battle.  OE tries to stick a bunch of environment info
into the build but tboot is pretty much self contained.
tboot squashes CFLAGS itself but we've gotta squash LDFLAGS for it.
Further the OE CPP gives tboot problems (it's really gcc with some
options) so we squash this and use the actual ${HOST_PREFIX}cpp
with toolchain options.

We have to patch the root Makefile to make it friendly to
cross compiling.

We include a COMPATIBLE_HOST of 32 and 64 bit x86 on linux and BSD
which I think is sufficient to limit tboot to compatible system.
We don't check for MACHINE types that support TXT, just HOSTs that
will execute 32bit x86 which is what the tboot binary is.  If the
system doesn't support TXT tboot should just pass with default
policy.

I'd rather be building this with '-C tboot' but for some reason
this prevents tboot from setting up it's variables properly so we
just squash the SUBDIRS from the root of the build.