Add recipe for sugarbay ACM.
Need a structured way to pull in the proper ACM for a distro using
tboot.  Likely a distro feature is the right way?  Basically all
that's needed is detecting the MACHINE and lining that up with the
proper ACM ... haven't though much about it yet as all I've got
are sugarbay systems.