Add initrd script that sets up the TPM stuff.
This init script is intended to be an example of the minimum work
necessary to get a shell in the initrd where a user can run tpm_*
commands.  It doesn't pivot to a rootfs or anything fancy.  Just
sets up the environment necessary for tcsd to startup then it kicks
off a shell.