Replace prototype syslinux multiboot config builder with static config.
I've moved the syslinux multiboot config builder to a topic branch in
an openembedded-core mirror on my github.  That work turned out to be
a bit of a diversion currently.  It's necessary To get the multiboot
config stuff working across both syslinux and grub.  The later for
EFI booting.  Since tboot doesn't support EFI boot yet this is low
priority.