measured-image-bootimg: Add grub-efi bits.
Append to grubefi_populate function to install the modules needed by
tboot. This is multiboot2, relocator and gzio. The acpi module is
a dependency of multiboot2.

Re-define the grubefi_iso_populate. This is nearly identical to the
original from the grub-efi.bbclass but copying the EFIDIR from iso_dir
to the EFIIMGDIR is now recursive to copy the modules directory.
This should go upstream.

Re-define the build_grub_cfg function. The original in grub-efi.bbclass
is nice and all but it doesn't support multiboot configurations.