trousers: Delete empty installed directory in do_install_append.
The trousers build creates an empty directory in ${libdir}/trousers
in the install step. As additional QA tests have been enabled in OE
our previous approach of just ignoring this directory has resulted
in a WARNING message. To quiet this we simply remove the empty
directory at the end of the do_install step.

Signed-off-by: Philip Tricca <flihp@twobit.us>