tboot: Update hashes for 1.9.4 release tarball.
The hashes for the 1.9.4 release tarball have changed because (AFAIK)
the tboot team released a new 1.9.4 tarball. The first release included
the .hg directory (mercurial SCM metadata) which bloats the tarball
significantly. This is just bad release discipline on the part of the
tboot project.

These sorts of changes are obnoxious for OE. You will only see this
failure if you're downloading the new release tarball / not working from
a donwload cache. OE will only pull down a new tarball if the version
number / URL changes (I think).

The "right way" to cope with changes like this (fixing a botched release
tarball) is for the upstream to bump the version number. This is a bug
in the 1.9.4 tboot release and so going to 1.9.5 for a fix is very
reasonable. I'm not holding out much hope for upstream to do this sort
of thing though.

Signed-off-by: Philip Tricca <flihp@twobit.us>