linux-intel: Add bbappend for new kernel recipe from meta-intel.
The meta-intel layer now sets PREFERRED_PROVIDER_virtual/kernel to
'linux-intel', which looks to be a variant of linux-yocto with some
intel specific stuff turned on. Depending on what BSP you're using this
may not effect you. But for BSPs derived from
conf/machine/include/meta-intel.inc in meta-intel you'll now be getting
this kernel instead of linux-yocto.

This patch adds a bbappend for this kernel so intel systems can still
get the TPM(2) MACHINE_FEATURE.

Signed-off-by: Philip Tricca <flihp@twobit.us>