tpm2.0: simplify autotools dependencies and setup
Normally, m4 macros are taken from -native recipes because that avoids
building otherwise unused packages for the target system.

We can make that work also for tpm2.0-tss by disabling the actual
autotools bootstrapping, i.e. bootstrap is only called to create the
file lists.  The custom bootstrapping no longer seems to be needed in
tpm2.0-tools.

"inherit pkgconfig" already ensures that pkgconfig-native is present,
so we do not need an explicit DEPENDS for it.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>