Remove rng-tools from core-image-tpm.inc
This is causing the image to fail when booting. The likely cause is the
rng-tool daemon starting before the tpm daemon and hanging waiting for
it to start. In the interest of getting a functional build we're
removing the rng-tools dependency. It can be added back and debugged
after the image is fixed.

Signed-off-by: Philip Tricca <flihp@twobit.org>