Split MACHINE_FEATURES up into tpm, tpm2-tis and tpm2-crb.
The previous implementation assumed that tpm2 implementations used the
crb interface only. The qemu tpm2 implementation exposes a tpm2 device
using the TIS interface. To support this we now have 3 MACHINE_FEATURES:
- tpm - A TPM 1.2 deivce using the TIS interface
- tpm2-tis - A TPM 2.0 device using the TIS interface
- tpm2-crb - A TPM 2.0 device using the CRB interface

The kernel config fragments have a 1:1 correspondence to these machine
features as do the packagegroups that pull in the required drivers and
software stack. Additionally the kernel config fragment has been updated
to build the TIS driver as a module.

Signed-off-by: Philip Tricca <flihp@twobit.org>