tpm2-tss: upgrade recipe from 2.1.0 to 2.3.1
This adds the new tcti loading library (libtss2-tctildr) as well as the
response code decoding library (libtss2-rc). The default crypto library
was changed upstream from gcrypt to openssl so this dependency is
changed in the bitbake metadata as well. Providing a mechanism for
chosing the crypto library is future work.

We must now also explicitly disable the generation of documentation by
passing the configure script the '--disable-doxygen-doc' option since
the configure script now reports an error if doxygen is not present.

Signed-off-by: Philip Tricca <flihp@twobit.org>