tboot: update recipes from 1.9.6 to 1.9.10
This required rebasing the tboot-cross-compile patch. GCC got a bit more
strict in the version 9 release (used by OE core) and it started
complaining about tboot extracting pointers from packed structures
(which can result in unaligned pointers). This required a new patch to
ignore the error message. A better fix upstream would be nice but
it was reported to them ~6 months ago but it's still an issue.

Signed-off-by: Philip Tricca <flihp@twobit.org>