wic-image-tpm: Add recipe to build wic image for rootfs wit tpm stuff.
wic is nice. Using the 'wic' executable to generate the rootfs is great
but integrating the whole thing with bitbake is better. This is a simple
example that creates a GPT partitioned disk that boots using grub-efi.
The rootfs comes from from core-image-tpm with a kernel config to boot
it readonly. There's a bit of swap space too (for whatever that's
worth).

Signed-off-by: Philip Tricca <flihp@twobit.org>