wic-image-tpm: Add IMAGE_TYPEDEP on ext4 for wic fstype.
Machine types from meta-intel (or based on them) require the user to
manually add this to their local.conf (per the instructions in
meta-intel). QEMU machine types from oe-core do not require this. Seems
more natural to have this in the wic-image-tpm recipe instead since the
wks.in file used ext4 directly.

Signed-off-by: Philip Tricca <flihp@twobit.org>